week_4/requirements.txt
supabase
python-dotenv
pymupdf
pdf2image


